Add ‘Switcher’ & ‘Message’ documentation articles
