+++ Sources/Atem/Documentation.docc/Messages.md
@@ -1,0 +1,9 @@
+# ``Atem/Message``
+@Metadata {
+	@DocumentationExtension(mergeBehavior: append)
+}
+
+Messages are split up in three categories:
+- ``Do`` describes **commands** to **perform** certain **actions**
+- ``Did`` describes **notifications** of performed actions
+- ``Config`` describes **configurations**
