+++ Sources/Atem/Messages/Message.swift
@@ -8,6 +8,14 @@
 
 import Foundation
 
+/// A shortcut to ``Message/Do``
+public typealias Do  = Message.Do
+/// A shortcut to ``Message/Did``
+public typealias Did = Message.Did
+/// A shortcut to ``Message/Config``
+public typealias Config = Message.Config
+
+/// A namespace for messages that are sent between ``Controller``s and ``Switcher``s
 public enum Message {
 	/// An ASCII String of exactly 4 characters. A list of available titles can be found in [Skarhoj's protocol description](http://skaarhoj.com/fileadmin/BMDPROTOCOL.html) under the column "CMD".
 	public struct Title: CustomStringConvertible {
@@ -70,10 +78,6 @@ public enum Message {
 	/// Namespace for configuration messages
 	public enum Config {}
 }
-
-public typealias Do  = Message.Do
-public typealias Did = Message.Did
-public typealias Config = Message.Config
 
 /// A message that can be constructed from a sequence of bytes (`ArraySlice<UInt8>`)
 public protocol DeserializableMessage: CustomDebugStringConvertible {
